aluguel_imoveis/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1.1
django-allauth==65.2.0
django-bootstrap4==24.4
django-bootstrap5==24.3
django-stdimage==6.0.2
pillow==10.4.0
pip-autoremove==0.10.0
psycopg2-binary==2.9.10
setuptools==75.6.0
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
